MOVEONLY: Move balance and utxo RPCs to coins.cpp
