+++ src/Makefile.am
@@ -412,6 +412,7 @@ libbitcoin_wallet_a_SOURCES = \
   wallet/receive.cpp \
   wallet/rpc/addresses.cpp \
   wallet/rpc/backup.cpp \
+  wallet/rpc/coins.cpp \
   wallet/rpc/encrypt.cpp \
   wallet/rpc/signmessage.cpp \
   wallet/rpc/transactions.cpp \
